requirements_backup.txt
# Dependencies for procedure.py
numpy
pyvista
vtk
trimesh
ezdxf
Pillow
svgwrite
scipy
networkx
svgpathtools